Formatting & Movement Limitations
Players are now limited to move only to appropriate spaces, and if a player tries to move to an invalid space their turn is skipped.
Locked game to 16:9 aspect ration to allow for better formatting of UI that persists in builds. Arranged current GameObjects to allow for the future addition of the player stacks.